:T3C-791 Improve Topic and Subtopic Descriptions
For Topics and Subtopics wither very few related comments, the descriptions are coming in longer than we would like.  We don't want the topic short description to be longer than the length of all the comments in the topic.  The same goes for the short description for each subtopic.

Run the clustering evaluation from @common by running `npm run eval:clustring` so that we get a baseline score

Update the scorer that checks shortDescription length in common/evaluations/clustering/scorers.ts to test based on the new logic.

If new data needs to be added to test against this logic update add that to common/evaluations/clustering/datasets.ts

Run the clustering evaluate from the @common directory by running `npm run eval:clustering` so that we can see the scores went down with our test changes.

Modify the defaultClusteringPrompt in common/prompts/index.ts to get the functionality we want.

Run the evaluation again to see if the score improved

Continue modifying the prompt and running the evaluation score until the score is equal or better than the original score before we modified anything.